refactor: consolidate user cube editing functionality into userCubeCard component
- Removed the editUserCubeCard component and integrated its functionality into userCubeCard, streamlining the user interface for cube management.
- Updated userCubeCard to support both view and edit modes, enhancing usability and reducing code duplication.
- Adjusted data binding and state management for user details, ensuring accurate updates and improved user experience during cube editing.
